+++ include/SubgradientSolution.hpp
@@ -11,17 +11,17 @@ class SubgradientSolution {
 	vector<pair<int, int>> edges;
 	vector<vector<pair<int, int>>> adjacencyList;
 	int higherDegreeVertex;
-	vector<double> penalizer;
-	double penalizerCost;
+	vector<int> penalizer;
+	int penalizerCost;
 
    public:
 	SubgradientSolution(double lb, vector<pair<int, int>> edges, int dimension);
 	shared_ptr<vector<pair<int, int>>> getForbiddenArcs();
 	vector<pair<int, int>> getEdges() { return edges; }
 	double getLowerBound() { return lb; }
-	vector<double> getPenalizer() { return penalizer; }
+	vector<int> getPenalizer() { return penalizer; }
 	bool isSolutionFeasible(shared_ptr<vector<double>> lambda);
-	double getPenalizerCost() { return penalizerCost; }
+	int getPenalizerCost() { return penalizerCost; }
 };
 
 #endif  // !SUBGRADIENTSOLUTION_HPP
